Version2_Work
Masomenos funciona ya hay una mayor interaccion entre ventanas. Sin embargo no se aprobado.
