+++ .gitignore
@@ -69,7 +69,7 @@ crashlytics.properties
 crashlytics-build.properties
 fabric.properties
 
-# Editor-based Rest Client
+# Editor-based Rest ClientObject
 .idea/httpRequests
 
 # Android studio 3.1+ serialized cache file
