+++ src/client/Client.java
@@ -1,64 +1,122 @@
 package client;
 
-import java.io.DataOutputStream;
+import javafx.geometry.Pos;
+import javafx.scene.*;
+import javafx.scene.control.*;
+import javafx.scene.layout.*;
+import javafx.stage.Stage;
+
 import java.io.IOException;
 import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
+import java.net.ServerSocket;
 import java.net.Socket;
 
-class Client {
-    private String nick,ip, text;
-    int puerto;
 
-    public Client(String nick, String ip, int puerto) {
-        this.nick = nick;
-        this.ip = ip;
-        this.puerto = puerto;
+public class Client implements Runnable {
+
+    private  Stage primaryStage = new Stage();
+    private Label labelTitle = new Label("ClientObject Chat");
+    private TextArea message = new TextArea();
+    private TextField input = new TextField();
+    private Button buttonSave = new Button("Save");
+    private Button buttonSend = new Button("send");
+    private TextField textFieldNick, textFieldIp, textFieldPuerto;
+    private ClientObject theClientObject = new ClientObject("Nickname","127.0.0.1", 6000);
+
+    public Client(String name, String ip, int puerto){
+        this.theClientObject = new ClientObject(name, ip, puerto);
+        display();
+        new Thread(this::run).start();
     }
 
 
-    public void sendMessageAction(String text){
-        this.text = text;
+
+
+    public void display() {
+       // Parent root = FXMLLoader.load(getClass().getResource("client.fxml"));
+
+//        Label
+        labelTitle.setStyle("-fx-text-fill: #1fdece; -fx-font-size: 36px; -fx-font-family: Arial");
+
+
+//       TextArea
+        message.setPrefHeight(150);
+        message.setStyle("-fx-font-family: Arial;" +
+                "-fx-font-size: 14px; " + "-fx-control-inner-background:#373737");
+        message.setEditable(false);
+
+
+//       TextField
+        input.setPrefWidth(350);
+
+        textFieldNick = new TextField(theClientObject.getNick());
+        textFieldNick.setPromptText("nickname");
+        textFieldNick.setDisable(true);
+
+        textFieldIp = new TextField(theClientObject.getIp());
+        textFieldIp.setPromptText("your ip");
+
+        textFieldPuerto = new TextField(String.valueOf(theClientObject.getPuerto()));
+        textFieldPuerto.setPromptText("puerto");
+
+//        Button
+        buttonSend.setPrefWidth(50);
+        buttonSend.setOnAction(event -> {
+            ControllerClient.sendMessage(theClientObject, input, message);
+        });
+
+        buttonSave.setOnAction(event -> {
+            ControllerClient.saveNewConfig(theClientObject, textFieldIp, textFieldPuerto);
+        });
+
+//        HBox
+        HBox hBox1 = new HBox();
+        hBox1.getChildren().addAll(textFieldNick, textFieldIp, textFieldPuerto);
+
+        HBox hBox2 = new HBox();
+        hBox2.getChildren().addAll(input, buttonSend);
+
+
+
+//        VBox
+        VBox root = new VBox(40);
+        root.setStyle("-fx-background-color:#282C34");
+        root.setPrefSize(400, 400);
+        root.getChildren().addAll(labelTitle, hBox1, buttonSave, message, hBox2);
+        root.setAlignment(Pos.TOP_CENTER);
+
+//        Scene
+        primaryStage.setTitle("ClientObject");
+        Scene scene = new Scene(root);
+
+//        Stage
+        primaryStage.setScene(scene);
+        primaryStage.setResizable(false);
+        primaryStage.show();
+
+
+    }
+
+
+    @Override
+    public void run() {
         try {
-            Socket socket = new Socket(ip, puerto);
+            ServerSocket serverSocket =  new ServerSocket( theClientObject.getPuerto() );
+            Socket socket;
+            while (true){
+                socket = serverSocket.accept();
 
-            ObjectOutputStream paquete =  new ObjectOutputStream(socket.getOutputStream());
+                ObjectInputStream objectInputStream = new ObjectInputStream(socket.getInputStream());
 
-            paquete.writeObject(this);
+                ClientObject clientObjectWriter = (ClientObject) objectInputStream.getObjectInputFilter();
 
-            socket.close();
+                this.message.appendText(clientObjectWriter.getNick() + ": " + clientObjectWriter.getText() + "\n" );
 
+                socket.close();
+            }
         } catch (IOException e) {
             e.printStackTrace();
         }
 
     }
-
-
-    public void setIp(String ip) {
-        this.ip = ip;
-    }
-
-    public void setPuerto(int puerto) {
-        this.puerto = puerto;
-    }
-
-    public String getIp() {
-        return ip;
-    }
-
-    public int getPuerto() {
-        return puerto;
-    }
-
-    public String getNick() {
-        return nick;
-    }
-
-    public String getText() {
-        String oldtext = this.text;
-        //this.text = "";
-        return oldtext;
-    }
 }
-
