+++ src/sample1/MainController.java
@@ -1,0 +1,15 @@
+package sample1;
+
+
+import javafx.event.ActionEvent;
+
+
+public class MainController {
+
+    public void openClient(ActionEvent event) {
+
+    }
+
+    public void  openServer(ActionEvent event){
+    }
+}
