+++ src/sample1/MainView.fxml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="250.0" prefWidth="426.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample1.MainController">
+   <Button layoutX="187.0" layoutY="70.0" mnemonicParsing="false" onAction="#openServer" text="Server" />
+   <Button layoutX="187.0" layoutY="152.0" mnemonicParsing="false" onAction="#openClient" text="Clients" />
+</AnchorPane>
